Refactor hunk handling logic and update imports for improved clarity and structure; remove obsolete resolveHunkReferencesInComments function and tests
