--- pr-review/src/domain/services/resolveHunkReferencesInComments.spec.ts
@@ -1,11 +1,0 @@
-// import { describe, expect, it } from "vitest"
-// import { resolveHunkReferencesInComments } from "./resolveHunkReferencesInComments.js"
-
-// describe("resolveHunkReferencesInComments", () => {
-//   it("returns empty array if no files match", () => {
-//     const comments = [{ path: "foo.js", start: 1, end: 1, comment: "test" }]
-//     const files: any[] = []
-//     const result = resolveHunkReferencesInComments(comments, files)
-//     expect(result).toEqual([])
-//   })
-// })
